Added ipynb2rst.py script to buildDoc.
The script is used to convert an IPython notebook into a RestructuredText file,
remove all IPython prompts and correct the generated image paths.

Call it like this:

    $ python ipynb2rst.py path/to/my/IPython/notebook.ipynb

If this proves acceptable, we can extend the script to simply process every
IPython notebook in the new buildDoc/ipythonNotebooks directory, and move their
generated image folders and rst files over into the buildDoc/rst directory.

That can then be hooked directly into the doc building script.

NOTE: for now, nbconvert.py needs to be available on your PATH for this to work.

It sounds like the developers of nbconvert.py intend that the tool be available
as an actual Python import, but that functionality isn't yet ready.